% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oTree.R
\docType{data}
\name{oTree}
\alias{oTree}
\title{Sample experiment data}
\format{
A list of data frames created by import_otree().
}
\source{
The data set was created by using modified versions of the official
oTree sample experiments that can be downloaded when installing oTree.
In detail, the following apps were used: 
"start," "dictator," "chatapp," "survey."
}
\usage{
oTree
}
\description{
Sample experiment data
}
\keyword{datasets}
